+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/BallEntry.java
@@ -2,11 +2,11 @@ package org.firstinspires.ftc.teamcode.common;
 
 public class BallEntry {
     public int index;
-    public int intakePosition;
-    public int launchPosition;
+    public double intakePosition;
+    public double launchPosition;
     public GameColors ballColor;
 
-    public BallEntry(int index, int intakePosition, int launchPosition, GameColors ballColor) {
+    public BallEntry(int index, double intakePosition, double launchPosition, GameColors ballColor) {
         this.index = index;
         this.intakePosition = intakePosition;
         this.launchPosition = launchPosition;
